Move methods from abstract class to the classes where they are used
There is no natural inheritance here, so I'm unwinding the inheritance in the code.  No methods were changed, they were just moved to different classes
